+++ package.json
@@ -99,11 +99,8 @@
     "mini-css-extract-plugin": "^0.4.1",
     "mkdirp": "0.5.1",
     "nightmare": "3.0.1",
-    "postcss-color-rgba-fallback": "^3.0.0",
-    "postcss-initial": "^2.0.0",
     "postcss-loader": "^2.1.6",
     "postcss-nested": "^3.0.0",
-    "postcss-pseudoelements": "^5.0.0",
     "postcss-simple-vars": "^4.1.0",
     "pretty-exceptions": "^1.4.0",
     "probe-image-size": "4.0.0",
@@ -122,6 +119,7 @@
     "root-path": "^0.2.1",
     "serve-handler": "3.3.1",
     "simple-git": "1.96.0",
+    "style-loader": "^0.21.0",
     "throttle-debounce": "2.0.0",
     "transform-markdown-links": "^2.0.0",
     "uglifyjs-webpack-plugin": "1.2.7",
