+++ webpack.config.js
@@ -1,124 +1,127 @@
-/* eslint-disable import/no-unresolved */
-
+const fs = require('fs');
 const path = require('path');
+const globby = require('globby');
+
 const webpack = require('webpack');
+const FriendlyErrorsWebpackPlugin = require('friendly-errors-webpack-plugin');
+const webpackMerge = require('webpack-merge');
 
-const WebpackCdnPlugin = require('webpack-cdn-plugin');
-const CopyWebpackPlugin = require('copy-webpack-plugin');
-const { BundleAnalyzerPlugin } = require('webpack-bundle-analyzer');
-const UglifyJsPlugin = require('uglifyjs-webpack-plugin');
-const HtmlWebPackPlugin = require('html-webpack-plugin');
-const CleanWebpackPlugin = require('clean-webpack-plugin');
-const merge = require('webpack-merge');
-const baseConfig = require('./webpack.config.base.js');
+const defaults = require('./dist/default-config');
+const markdownItConfig = require('./markdownit.config');
+const IgnitePlugin = require('./dist/plugins/IgniteInjectPlugin');
+const babelRc = require('./.babelrc');
 
-module.exports = function(options) {
-  process.env.BABEL_ENV = 'production';
+const babelConfig = options =>
+  babelRc({
+    env() {
+      return options.mode;
+    }
+  });
+const modeConfig = env =>
+  require(`./build-utils/webpack.config.${env.mode}`)(env);
+// const presetConfig = require("./build-utils/loadPresets");
 
+module.exports = options => {
+  options = {
+    ...defaults,
+    ...options
+  };
+
+  const { ignitePlugins } = markdownItConfig.splitPlugins(options.plugins);
+
+  const docs = globby.sync([path.join(options.src, '**/*.md')]);
+  const logoPath = options.logo ? path.join(options.src, options.logo) : null;
+  const logoExists = logoPath && fs.existsSync(path.resolve(logoPath));
   const dest = options.dst ? path.resolve(options.dst) : null;
 
-  return merge(baseConfig(options), {
-    profile: options.json,
+  return webpackMerge(
+    {
+      mode: options.mode,
 
-    optimization: {
-      minimizer: [
-        new UglifyJsPlugin({
-          uglifyOptions: {
-            parse: {
-              // We want uglify-js to parse ecma 8 code. However, we don't want it
-              // to apply any minfication steps that turns valid ecma 5 code
-              // into invalid ecma 5 code. This is why the 'compress' and 'output'
-              // sections only apply transformations that are ecma 5 safe
-              // https://github.com/facebook/create-react-app/pull/4234
-              ecma: 8
-            },
-            compress: {
-              ecma: 5,
-              warnings: false,
-              // Disabled because of an issue with Uglify breaking seemingly valid code:
-              // https://github.com/facebook/create-react-app/issues/2376
-              // Pending further investigation:
-              // https://github.com/mishoo/UglifyJS2/issues/2011
-              comparisons: false
-            },
-            minify: {
-              toplevel: true
-            },
-            mangle: {
-              safari10: true
-            },
-            output: {
-              ecma: 5,
-              comments: false,
-              // Turned on because emoji and regex is not minified properly using default
-              // https://github.com/facebook/create-react-app/issues/2488
-              // eslint-disable-next-line camelcase
-              ascii_only: true
-            }
-          },
-          // Use multi-process parallel running to improve the build speed
-          // Default number of concurrent runs: os.cpus().length - 1
-          parallel: true,
-          // Enable file caching
-          cache: true,
-          sourceMap: true
-        })
-      ]
-    },
+      devtool: 'source-map',
 
-    plugins: [
-      new webpack.PrefetchPlugin('', 'react-ideal-image'),
-      new webpack.PrefetchPlugin('', 'react-color'),
-      new webpack.PrefetchPlugin('', 'react-waypoint'),
-      new CleanWebpackPlugin([dest]),
-      new HtmlWebPackPlugin({
-        base: options.baseURL,
-        codeStyle: options.codeStyle,
-        bulmaTheme: options.bulmaTheme,
-        template: path.resolve(__dirname, './src/index.html'),
-        filename: './index.html',
-        minify: {
-          removeComments: true,
-          collapseWhitespace: true,
-          removeRedundantAttributes: true,
-          useShortDoctype: true,
-          removeEmptyAttributes: true,
-          removeStyleLinkTypeAttributes: true,
-          keepClosingSlash: true,
-          minifyJS: true,
-          minifyCSS: true,
-          minifyURLs: true
-        }
-      }),
-      new HtmlWebPackPlugin({
-        baseURL: options.baseURL,
-        template: path.resolve(__dirname, './src/404.html'),
-        filename: './404.html'
-      }),
-      options.static &&
-        new CopyWebpackPlugin([
-          {
-            from: path.join(options.src, '**/*.{jpg,png,gif}')
-          }
-        ]),
-      new WebpackCdnPlugin({
-        modules: {
-          react: [
-            {
-              name: 'react',
-              var: 'React',
-              path: `umd/react.${options.mode}.min.js`
-            },
-            {
-              name: 'react-dom',
-              var: 'ReactDOM',
-              path: `umd/react-dom.${options.mode}.min.js`
-            }
-          ]
-        },
+      entry: [
+        logoExists ? path.resolve(logoPath) : null,
+        path.resolve(__dirname, './src/app/index.js')
+      ].filter(Boolean),
+
+      output: {
+        path: dest,
         publicPath: options.baseURL
-      }),
-      options.analyze && new BundleAnalyzerPlugin()
-    ].filter(Boolean)
-  });
+      },
+
+      resolve: {
+        alias: {
+          ignite: path.resolve(__dirname, './src/app/index.js')
+        }
+      },
+
+      module: {
+        rules: [
+          // Markdown
+          {
+            test: /\.md$/,
+            use: [
+              {
+                loader: 'babel-loader',
+                options: babelConfig(options)
+              },
+              {
+                loader: path.resolve(
+                  __dirname,
+                  './dist/loaders/html-to-react.js'
+                ),
+                options
+              },
+              {
+                loader: 'markdown-it-vanilla-loader',
+                options: markdownItConfig.generateConfig(options.plugins)
+              },
+              {
+                loader: path.resolve(__dirname, './dist/loaders/hash-link.js'),
+                options
+              }
+            ]
+          },
+          // Javascript
+          {
+            test: /\.js$/,
+            exclude: /node_modules\/(?!.*ignite\/src)/,
+            use: {
+              loader: 'babel-loader',
+              options: babelConfig(options)
+            }
+          }
+        ]
+      },
+
+      plugins: [
+        new IgnitePlugin({
+          entries: docs.map(doc => path.resolve(doc)),
+          plugins: ignitePlugins,
+          options
+        }),
+        new webpack.DefinePlugin({
+          'process.env': {
+            SEARCH: JSON.stringify(options.searchIndex),
+            index: JSON.stringify(options.index),
+            baseURL: JSON.stringify(options.baseURL),
+            title: JSON.stringify(options.title),
+            githubURL: JSON.stringify(options.githubURL),
+            navItems: JSON.stringify(options.navItems),
+            logo: JSON.stringify(logoExists ? logoPath : ''),
+            plugins: JSON.stringify(options.plugins)
+          }
+        }),
+        options.log &&
+          new FriendlyErrorsWebpackPlugin({
+            clearConsole: options.mode !== 'production',
+            compilationSuccessInfo: options.compilationSuccessInfo
+          }),
+        ...options.webpackPlugins
+      ].filter(Boolean)
+    },
+    modeConfig(options)
+    // presetConfig({ mode, presets })
+  );
 };
